A few more...

+++ trunk/namespaceobj.c
@@ -203,7 +203,7 @@ MUST_CHECK Obj *namespace_member(oper_t op, Namespace *v1) {
             if (count >= 0) count++;
             err = new_error(ERROR___NOT_DEFINED, &v2->epoint);
             err->u.notdef.names = ref_namespace(v1);
-            err->u.notdef.ident.len = count;
+            err->u.notdef.ident.len = (size_t)count;
             err->u.notdef.ident.data = NULL;
             err->u.notdef.down = false;
             return &err->v;
